enemy and govno

+++ Assets/Scripts/Factories/Struct/EnemyStats.cs
@@ -3,6 +3,6 @@ public struct EnemyStats
 {
     public int Damage;
     public int Health;
-    public int MoveSpeed;
+    public float MoveSpeed;
     public int Power;
 }
